Move podNIC related code to podnic.go
